mudado a tipagem dos usuários para pasta model
